Created the Spawn System
In addition of spawning the character on scene changes + with a publicly callable SpawnCharacter method for runtime spawnings, I also implemented an event system which gets invoked after a character is spawned in scene. The event takes the GameObject that was spawned as a parameter for furthermore runtime checks. This is an utility that I use on my own personal projects to check from my game manager which player is spawned, do the spawned player has any specific components, tags etc than act accordingly.