earring for townsfolk
Added an Earring for townsfolk. It was build with Probuilder, exported to obj for some finishing touches with Blender, then saved as obj again. The material is a simple color.

The earring possibly needs to be placed somewhere else once the townfolk is rigged. Which begs the question whether they can move their ears...