Game Settings and preview implementation
Game Settings makes it easy to create setting and edit UI without writing additional logic! Check out the OptionsMenu - Preview scene for this.

As a test, try adding a Shadow Quality Dropdown to the scene and have it display the "Hard Only" option.

Support for URP and the Input System is coming.