Fixed colliders on the trees
Added child GO to the trees named "leaves", added colliders to the leaves and set layer of leaves to "TreeLeaves" (So, also added this new layer. Set in physics properties to not collide with anything). In CameraSystem/FreeLook_VCam in "Cinemachine collider" added "TreeLeaves" layer in the field "Collide Against".