Oven from DanTen66
Here is the Oven that I have been working on with Albedo, Material and a Prefab Variant. I am including the .meta files just in case.

You mentioned not to adjust the material shader, though this one has an outline and from what I understand the buildings and enviroment are not supposed to.