Fading to black on scene change
Created ScreenFaderToBlack MonoBehaviour which can be attatched to the UI element with Image component.

Functions are called through BoolEventChannelSO.
Assign apropriate SpriptableObject to the ScreenFaderToBlack MonoBehaviour and To the caller script.  In caller script just call "RaiseEvent(true/false)" method of Scriptable Object.

Implemented fading on scene loading in LocationLoader.

Deleted Debug.Log's in scripts (LocationLoader, CameraManager)

In ScreenFaderToBlack  Script can be adjusted time fade to black and fade from black.
Handles situations like - toBlack -> fromBlack -> toBlack - so that step is executed to the end and only then a new step starts.

Small bug fix: added PopulateSceneList() in SceneAccessTool in OnEnable because without it scenes list is not initialised when window is opened.