Basic structure for saving location to a file
Added a Uuid type and drawer

Removed saving TestScript

Added a Database and Desc type to reference other SOs by UUID

When it comes to serializing data (scenes and items for example) it's much simpler to be able to reference everything by a UUID and to be able to easily look them up in a table.

Added a uuid to the LocationSOs and put them in a DB

Converted SaveSystem into a SO. Removed test vars from Save.