Adding a menu item that automatically generates a Tags class so we can use constant defined strings for tag comparison in the scripts.
The generator is pre-configured to generate a file called "Tags.cs". It contains a sealed class called "Tags" which resides in the "UOP1" namespace. This class contains "public const strings" of each tag in the project.

There are protections in the generator to stop accidental overwriting of already existing files in the project should the configuration values change.

The generated file Tags file should be committed to the repository (and is included in this commit) to save others from regenerating it.

Finally, it contains instructions on how to regenerate the class when a tag is added or removed from the project.