Adding a menu item that automatically generates a Layer and LayerMasks enums so we can use their defined values for layer comparison in the scripts.
The generator is pre-configured to generate a file called "Layers.cs". It contains a pair of enums; Layer and LayerMasks. These reside in the "UOP1" namespace. These enums contains the layer ID's in the project and their respective layer masks. The LayerMasks enum has the Flags attribute applied to allow you to correctly use these LayerMasks to check against multiple layers at once, such as in Physics.Raycast.

There are protections in the generator to stop accidental overwriting of already existing files in the project should the configuration values change.

The generated Layer file should be committed to the repository (and is included in this commit) to save others from regenerating it. The generated file also contains instructions on how to use it and how to regenerate the file should a new layer be added to the project.

Finally, I included a handful of extension methods for UnityEngine.LayerMask to make it easy to check if a layer is contained within a UnityEngine.LayerMask.