Using the EditorApplication.projectChanged the LayerEnumsGenerator now listens for changes in the project, checks if the layers have changed and updates Layers.cs automatically. This should make it simpler and less error prone when adding new layers to the project, as the file should generate itself.
Sometimes this takes a few seconds as messages from EditorApplication.projectChanged don't happen immediately. Often it requires you to open a new folder or touch an asset in the project window to trigger it.

I've also left a menu option to regenerating the Layers.cs and removed the "FileExists" check.