Removed the tight dependency on the Tag type in the generator itself. The generator now searches for the type in the namespace and assembly defined in the configuration region.
The generator itself should not contain compilation errors no matter the configuration settings. However, expect compilation errors elsewhere in the project once the Tag constants are in use and are subsequently changed in this generator. This is a good thing as it'll show at compile time where the bugs might have been hiding by using "magic strings" (essentially, the whole point of this tool existing in the first place).