Added Statue(Mesh, Material, and Prefabs)
Statue was a contribution that was discussed in the forums and not based on any codecks card. The name I do have for him is from a Youtube Video, so I don't have any ideas for an original name if needed.

Meshes and Prefabs are in their specific folders.

Please delete the questions below, and type the answers in their place instead